update:EntityTrim-relation(ManyToOne=>,{onDelete:'CASCADE'}

+++ src/modules/board/board.service.ts
@@ -17,6 +17,7 @@ import { Colors } from "src/common/custom/types/enum-color.type";
 import { MESSAGES } from "src/common/constants/messages.constant";
 import { CardsEntity } from "src/entities/cards.entity";
 import { ListsEntity } from "src/entities/lists.entity";
+import { UsersEntity } from "src/entities/users.entity";
 
 import { Redis } from "ioredis";
 
@@ -32,10 +33,32 @@ export class BoardService {
     private cardRepository: Repository<CardsEntity>,
     @InjectRepository(ListsEntity)
     private listRepository: Repository<ListsEntity>,
+    @InjectRepository(UsersEntity)
+    private UserRepository: Repository<UsersEntity>,
     @Inject("REDIS_CLIENT") private redisClient: Redis,
   ) {}
 
+  //convenience function section
+  async findMember(boardId: number, userId: number) {
+    const member = await this.memberRepository.findOne({ where: { userId, boardId } });
+    return member;
+  }
+
+  async checkMemberRole(boardId, userId: number, role: MemberRoles) {
+    const member = await this.findMember(boardId, userId);
+    if (member.role === role) {
+      return true;
+    }
+    return false;
+  }
+
+  //==============================
+
+  //보드 만들기
   async createBoard(createBoardDto: CreateBoardDto, userId: number) {
+    //authorization section
+
+    //====================================================
     const queryRunner = this.dataSource.createQueryRunner();
     await queryRunner.connect();
     await queryRunner.startTransaction();
@@ -46,6 +69,7 @@ export class BoardService {
         userId: userId,
         role: MemberRoles.ADMIN,
         nickname: createBoardDto.nickname,
+        isCreator: true,
       });
 
       // if(!transactionCreateBoard) throw new {}
@@ -60,7 +84,13 @@ export class BoardService {
     }
   }
 
-  async updateBoard(boardId: number, updateBoardDto: UpdateBoardDto) {
+  //보드 정보 업데이트
+  async updateBoard(boardId: number, updateBoardDto: UpdateBoardDto, userId: number) {
+    //authorization section
+    if (await this.checkMemberRole(+boardId, userId, MemberRoles.ONLY_VIEW))
+      //ONLY_VIEW일 경우 권한이 없다.
+      throw new ConflictException("접근 권한이 없습니다.");
+    //====================================================
     const board = await this.boardRepository.findOne({ where: { id: boardId } });
     await this.boardRepository.update(
       { id: boardId },
@@ -73,15 +103,29 @@ export class BoardService {
     return { Message: BOARD_CONSTANT.MODIFY_BOARD };
   }
 
-  async deleteBoard(boardId: string) {
-    await this.boardRepository.delete(boardId);
+  //보드 삭제
+  async deleteBoard(boardId: string, userId: number) {
+    //authorization section
+    if (await this.checkMemberRole(+boardId, userId, MemberRoles.ADMIN)) {
+      await this.boardRepository.delete(boardId);
+    }
+    throw new ConflictException("접근 권한이 없습니다.");
+    //ONLY_VIEW일 경우 권한이 없다.
   }
 
+  //보드 찾기
   async findBoard(boardId: number, userId: number) {
+    //authorization section
+    if (await this.checkMemberRole(+boardId, userId, MemberRoles.ONLY_VIEW))
+      //ONLY_VIEW일 경우 권한이 없다.
+      throw new ConflictException("접근 권한이 없습니다.");
+
     const member = await this.memberRepository.findOne({ where: { boardId, userId } });
     if (!member) {
       throw new ConflictException("접근 권한이 없습니다.");
     }
+    //====================================================
+
     const board = await this.boardRepository.findOne({ where: { id: boardId } });
     const lists = await this.listRepository.find({ where: { boardId } });
     const listIds = lists.map((list) => list.id);
@@ -102,10 +146,17 @@ export class BoardService {
     return boardData;
   }
 
-  async inviteBoard(boardId: string) {
+  //보드 초대 코드 만들기
+  async inviteBoard(boardId: string, userId: number) {
+    //authorization section
+    //1.역할에 따른 권한 분류
+    if (await this.checkMemberRole(+boardId, userId, MemberRoles.ONLY_VIEW))
+      //ONLY_VIEW일 경우 권한이 없다.
+      throw new ConflictException("접근 권한이 없습니다.");
+
     const data = await this.redisClient.set(boardId, `inviteLink/board$${boardId}`);
     const result = await this.redisClient.get(boardId);
 
-    return { message: BOARD_CONSTANT.MAKE_INVITECODE };
+    return result;
   }
 }
